-Add MoviesDetail and his components to DI Modules -Create GetMoviesDetailsSingleUseCase -Add Date extensions -Add SpannableString extensions

+++ app/src/main/java/br/com/flying/dutchman/data/cache/RoomRepository.kt
@@ -30,7 +30,7 @@ class RoomRepository @Inject constructor(
         }
     }
 
-    override fun getMovie(movieId: Long): Single<MovieEntity> {
+    override fun getMovie(movieId: Int): Single<MovieEntity> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
